pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "khorsyio"
version = "0.2.0"
description = "Async event-driven python framework: ASGI, socketio, msgspec, SQLAlchemy 2.0"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "khorsyioio", email = "info@khorsyio.io"},
]
keywords = ["asyncio", "asgi", "event-driven", "socketio", "sqlalchemy", "framework"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Framework :: AsyncIO",
]
dependencies = [
    "msgspec>=0.18.6",
    "python-socketio>=5.11.0",
    "asyncpg>=0.29.0",
    "SQLAlchemy>=2.0.25",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "granian>=1.0.0",
    "uvicorn>=0.30.0"
]

[project.urls]
Homepage = "https://github.com/khorsyioio/khorsyio"
Repository = "https://github.com/khorsyioio/khorsyio"
Issues = "https://github.com/khorsyioio/khorsyio/issues"

[project.optional-dependencies]
granian = ["granian>=1.0.0"]
uvicorn = ["uvicorn>=0.30.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["khorsyio*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
